+++ src/pretalx/orga/forms/mails.py
@@ -1,41 +1,85 @@
+import string
+from collections import defaultdict
+
 from django import forms
+from django.db import transaction
 from django.db.models import Q
 from django.utils.translation import gettext_lazy as _
 from i18nfield.forms import I18nModelForm
 
 from pretalx.common.mixins.forms import ReadOnlyFlag
-from pretalx.mail.context import get_context_explanation
+from pretalx.mail.context import get_available_placeholders
 from pretalx.mail.models import MailTemplate, QueuedMail
 from pretalx.person.models import User
 
 
-class MailTemplateForm(ReadOnlyFlag, I18nModelForm):
+class MailTemplateBase(I18nModelForm):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        available_placeholders = ", ".join(
+            [
+                "{" + f"{placeholder}" + "}"
+                for placeholder in self.get_valid_placeholders()
+            ]
+        )
+        self.fields["text"].help_text = (
+            _("Available placeholders:") + " " + available_placeholders
+        )
+        self.fields["subject"].help_text = (
+            _("Available placeholders:") + " " + available_placeholders
+        )
+
+    def _clean_for_placeholders(self, text, valid_placeholders):
+        cleaned_data = super().clean()
+        valid_placeholders = set(valid_placeholders)
+        # We need to do this here, since we need to be sure about the valid placeholders first
+        used_placeholders = set()
+        for field in ("subject", "text"):
+            for lang in cleaned_data[field].data.values():
+                used_placeholders |= set(
+                    [v[1] for v in string.Formatter().parse(lang) if v[1]]
+                )
+        return used_placeholders - valid_placeholders
+
+    class Meta:
+        model = MailTemplate
+        fields = ["subject", "text"]
+
+
+class MailTemplateForm(ReadOnlyFlag, MailTemplateBase):
     def __init__(self, *args, event=None, **kwargs):
         self.event = event
         if event:
             kwargs["locales"] = event.locales
         super().__init__(*args, **kwargs)
 
+    def get_valid_placeholders(self):
+        kwargs = ["event", "submission", "user", "slot"]
+        valid_placeholders = []
+
+        if self.instance and self.instance.id:
+            if self.instance == self.event.update_template:
+                valid_placeholders.append("notifications")
+                kwargs = ["event", "user"]
+            elif self.instance == self.event.question_template:
+                valid_placeholders.append("questions")
+                valid_placeholders.append("url")
+                kwargs = ["event", "user"]
+
+        valid_placeholders += list(
+            get_available_placeholders(event=self.event, kwargs=kwargs).keys()
+        )
+        return valid_placeholders
+
     def clean_text(self):
         text = self.cleaned_data["text"]
-        if self.instance and self.instance.id:
-            context = None
-            if self.instance == self.event.update_template:
-                context = {"notifications": "test", "event_name": "test"}
-            elif self.instance == self.event.question_template:
-                context = {"questions": "test", "event_name": "test", "url": "test"}
-            elif self.instance in self.instance.event.fixed_templates:
-                context = {item["name"]: "test" for item in get_context_explanation()}
-            if context:
-                try:
-                    for language, local_text in text.data.items():
-                        local_text.format(**context)
-                except KeyError as e:
-                    raise forms.ValidationError(
-                        ('Unknown template key: "{key}", locale: {locale}').format(
-                            key=e.args[0], locale=language
-                        )
-                    )
+        valid_placeholders = self.get_valid_placeholders()
+        warnings = self._clean_for_placeholders(text, valid_placeholders)
+        if warnings:
+            warnings = ", ".join("{" + w + "}" for w in warnings)
+            raise forms.ValidationError(
+                str(_("Unknown template key!")) + " " + warnings
+            )
         return text
 
     class Meta:
@@ -85,7 +129,7 @@ class MailDetailForm(ReadOnlyFlag, forms.ModelForm):
         widgets = {"to_users": forms.SelectMultiple(attrs={"class": "select2"})}
 
 
-class WriteMailForm(I18nModelForm):
+class WriteMailForm(MailTemplateBase):
     recipients = forms.MultipleChoiceField(
         label=_("Recipient groups"),
         choices=(
@@ -155,24 +199,32 @@ class WriteMailForm(I18nModelForm):
             self.fields["subject"].help_text = _(
                 "If you provide only one language, that language will be used for all emails. If you provide multiple languages, the best fit for each speaker will be used."
             )
-        self.fields["text"].help_text = _(
-            "Please note: Placeholders will not be substituted, this is an upcoming feature. "
-            "Leave no placeholders in this field."
-        )
+        self.warnings = []
+
+    def get_valid_placeholders(self):
+        kwargs = ["event", "submission", "user"]
+        if getattr(self, "cleaned_data", None):
+            recipients = self.cleaned_data.get("recipients")
+            if recipients and not set(recipients) - set(["accepted", "confirmed"]):
+                kwargs.append("slot")
+        else:  # Not cleaned yet, we only need this for the help text
+            kwargs.append("slot")
+        return get_available_placeholders(event=self.event, kwargs=kwargs)
 
     def get_recipient_submissions(self):
         submissions = self.event.submissions.all()
 
-        query = Q()
         submission_states = [
             s
             for s in ["submitted", "accepted", "confirmed", "rejected"]
-            if s in self.cleaned_data["recipeients"]
+            if s in self.cleaned_data["recipients"]
         ]
         if submission_states:
-            query = query | Q(state__in=submission_states)
+            query = Q(state__in=submission_states)
+        else:
+            query = Q(state__isnull=False)
 
-        if "no_slides":
+        if "no_slides" in self.cleaned_data["recipients"]:
             query = query | Q(resources__isnull=True)
 
         filter_submissions = self.cleaned_data.get("submissions")
@@ -188,12 +240,50 @@ class WriteMailForm(I18nModelForm):
         submission_types = self.cleaned_data.get("submission_types")
         if submission_types:
             submissions = submissions.filter(submission_type__in=submission_types)
-        return submissions
+        return submissions.select_related(
+            "track", "submission_type", "event"
+        ).prefetch_related("speakers")
 
     def clean(self):
-        # TODO validate placeholders
-        # - room/start/end are only allowed when all talks are in accepted/confirmed
-        return super().clean()
+        cleaned_data = super().clean()
+        valid_placeholders = self.get_valid_placeholders().keys()
+        self.warnings = self._clean_for_placeholders(
+            cleaned_data["subject"], valid_placeholders
+        ) | self._clean_for_placeholders(cleaned_data["text"], valid_placeholders)
+        return cleaned_data
+
+    @transaction.atomic
+    def save(self):
+        self.instance.event = self.event
+        self.instance.is_auto_created = True
+        template = super().save()
+
+        submissions = self.get_recipient_submissions()
+        mails_by_user = defaultdict(list)
+        result = []
+        # First, render all emails
+        for submission in submissions:
+            for speaker in submission.speakers.all():
+                mail = template.to_mail(
+                    user=None,
+                    event=self.event,
+                    locale=speaker.locale,
+                    context_kwargs={"submission": submission, "user": speaker},
+                    commit=False,
+                    allow_empty_address=True,
+                )
+                mails_by_user[speaker].append(mail)
+
+        for user, user_mails in mails_by_user.items():
+            # Second, deduplicate mails: we don't want speakers to receive the same
+            # email twice, just because they have multiple submissions.
+            mail_dict = {m.subject + m.text: m for m in user_mails}
+            # Now we can create the emails and add the speakers to them
+            for mail in mail_dict.values():
+                mail.save()
+                mail.to_users.add(user)
+                result.append(mail)
+        return result
 
     class Meta:
         model = MailTemplate